requirements.txt
Flask==3.0.0
qrcode==7.3.1
pillow==10.1.0
colorthief==0.2.1
